+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Fixed WriteSeq32/64 gate outputs when not running (they are now off but will give a pulse on every write or movement of the active step)
 - Disallow random sequence length of 1 in all Phrase/Gate sequencers, and Foundry
+- Added option for chaining velocity settings in TwelveKey via right-click menu
 
 
 ### 1.1.0 (2019-07-04)